final touches to python version
